# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Setspace Agency Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create or update a user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks
      parameters:
        - name: assigneeId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{taskId}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /quality-checks:
    get:
      operationId: listQualityChecks
      tags: [quality]
      summary: List quality check submissions
      parameters:
        - name: taskId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Quality checks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QualityCheck"
    post:
      operationId: createQualityCheck
      tags: [quality]
      summary: Create quality check
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQualityCheckBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QualityCheck"
  /quality-checks/{checkId}:
    patch:
      operationId: updateQualityCheck
      tags: [quality]
      summary: Update quality check
      parameters:
        - name: checkId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQualityCheckBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QualityCheck"
  /messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages (team chat)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: before
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: createMessage
      tags: [messages]
      summary: Post a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMessageBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
  /notifications/{notificationId}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        profileImage:
          type: string
          nullable: true
        role:
          type: string
        department:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, username, firstName, lastName, role, createdAt]
    CreateUserBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        department:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
      required: [firstName, lastName]
    UpdateUserBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        department:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
        priority:
          type: string
        assigneeId:
          type: string
          nullable: true
        assignee:
          $ref: "#/components/schemas/User"
          nullable: true
        createdById:
          type: string
        dueDate:
          type: string
          nullable: true
        completedAt:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, status, priority, createdById, createdAt, updatedAt]
    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
        priority:
          type: string
        assigneeId:
          type: string
          nullable: true
        dueDate:
          type: string
          nullable: true
      required: [title]
    UpdateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
        priority:
          type: string
        assigneeId:
          type: string
          nullable: true
        dueDate:
          type: string
          nullable: true
        completedAt:
          type: string
          nullable: true
...
    QualityCheck:
      type: object
      properties:
        id:
          type: integer
        taskId:
          type: integer
          nullable: true
        reviewerId:
          type: string
        reviewer:
          $ref: "#/components/schemas/User"
          nullable: true
        submitterId:
          type: string
        submitter:
          $ref: "#/components/schemas/User"
          nullable: true
        rating:
          type: integer
        feedback:
          type: string
          nullable: true
        status:
          type: string
        videoVersionId:
          type: integer
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, reviewerId, submitterId, rating, status, createdAt, updatedAt]
    CreateQualityCheckBody:
      type: object
      properties:
        taskId:
          type: integer
          nullable: true
        submitterId:
          type: string
        rating:
          type: integer
        feedback:
          type: string
          nullable: true
        status:
          type: string
        videoVersionId:
          type: integer
          nullable: true
      required: [submitterId, rating]
    UpdateQualityCheckBody:
      type: object
      properties:
        rating:
          type: integer
        feedback:
          type: string
          nullable: true
        status:
          type: string
    Message:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        authorId:
          type: string
        author:
          $ref: "#/components/schemas/User"
          nullable: true
        parentId:
          type: integer
          nullable: true
        createdAt:
          type: string
      required: [id, content, authorId, createdAt]
    CreateMessageBody:
      type: object
      properties:
        content:
          type: string
        parentId:
          type: integer
          nullable: true
      required: [content]
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        type:
          type: string
        title:
          type: string
        body:
          type: string
          nullable: true
        isRead:
          type: boolean
        linkUrl:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, userId, type, title, isRead, createdAt]
...