# lib/api-zod/src/generated/types/updateTaskBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Setspace Agency Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateTaskBody {
  title?: string;
  description?: string | null;
  status?: string;
  priority?: string;
  assigneeId?: string | null;
  dueDate?: string | null;
  completedAt?: string | null;
}